Move JaxClassifier to experimental
