+++ art/experimental/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains the experimental Estimator API.
+"""
+from art.experimental.estimators.jax import JaxEstimator
