+++ art/experimental/estimators/classification/__init__.py
@@ -1,0 +1,4 @@
+"""
+Experimental classifiers.
+"""
+from art.experimental.estimators.classification.jax import JaxClassifier
